+++ src/index.js
@@ -12,7 +12,7 @@ resolver.define('getIssues', async ({ payload }) => {
   const baseFields = [
     'summary', 'assignee', 'status', 'priority', 'duedate',
     'customfield_10015', 'labels', 'issuetype', 'project', 'resolution', 'reporter',
-    'issuelinks',
+    'issuelinks', 'parent', 'customfield_10014',
   ];
   const allFields = [...new Set([...baseFields, ...extraFields])].join(',');